+++ src/layouts/home-assistant-main.js
@@ -11,6 +11,7 @@ import '../util/ha-url-sync.js';
 import '../util/hass-mixins.js';
 import './partial-cards.js';
 import './partial-panel-resolver.js';
+import EventsMixin from '../mixins/events-mixin.js';
 
 import(/* webpackChunkName: "more-info-dialog" */ '../dialogs/ha-more-info-dialog.js');
 import(/* webpackChunkName: "voice-command-dialog" */ '../dialogs/ha-voice-command-dialog.js');
@@ -18,7 +19,7 @@ import(/* webpackChunkName: "voice-command-dialog" */ '../dialogs/ha-voice-comma
 {
   const NON_SWIPABLE_PANELS = ['kiosk', 'map'];
 
-  class HomeAssistantMain extends window.hassMixins.EventsMixin(PolymerElement) {
+  class HomeAssistantMain extends EventsMixin(PolymerElement) {
     static get template() {
       return html`
     <style>
