+++ src/panels/config/automation/ha-automation-editor.js
@@ -35,7 +35,7 @@ function AutomationEditor(mountEl, props, mergeEl) {
 
 /*
  * @appliesMixin window.hassMixins.LocalizeMixin
- * @appliesMixin window.hassMixins.EventsMixin
+ * @appliesMixin NavigateMixin
  */
 class HaAutomationEditor extends
   window.hassMixins.LocalizeMixin(window.hassMixins.NavigateMixin(PolymerElement)) {
