+++ src/panels/config/automation/ha-automation-picker.js
@@ -16,7 +16,7 @@ import '../ha-config-section.js';
 import computeStateName from '../../../common/entity/compute_state_name.js';
 /*
  * @appliesMixin window.hassMixins.LocalizeMixin
- * @appliesMixin window.hassMixins.EventsMixin
+ * @appliesMixin NavigateMixin
  */
 class HaAutomationPicker extends
   window.hassMixins.LocalizeMixin(window.hassMixins.NavigateMixin(PolymerElement)) {
