+++ src/panels/config/cloud/ha-config-cloud-forgot-password.js
@@ -6,9 +6,12 @@ import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 import '../../../components/buttons/ha-progress-button.js';
 import '../../../layouts/hass-subpage.js';
 import '../../../resources/ha-style.js';
-import '../../../util/hass-mixins.js';
+import EventsMixin from '../../../mixins/events-mixin.js';
 
-class HaConfigCloudForgotPassword extends window.hassMixins.EventsMixin(PolymerElement) {
+/*
+ * @appliesMixin EventsMixin
+ */
+class HaConfigCloudForgotPassword extends EventsMixin(PolymerElement) {
   static get template() {
     return html`
     <style include="iron-flex ha-style">
