+++ src/panels/config/customize/types/ha-customize-array.js
@@ -4,9 +4,12 @@ import '@polymer/paper-listbox/paper-listbox.js';
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
-import '../../../../util/hass-mixins.js';
+import EventsMixin from '../../../../mixins/events-mixin.js';
 
-class HaCustomizeArray extends window.hassMixins.EventsMixin(PolymerElement) {
+/*
+ * @appliesMixin EventsMixin
+ */
+class HaCustomizeArray extends EventsMixin(PolymerElement) {
   static get template() {
     return html`
     <style>
