+++ src/panels/dev-event/events-list.js
@@ -1,9 +1,12 @@
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
-import '../../util/hass-mixins.js';
+import EventsMixin from '../../mixins/events-mixin.js';
 
-class EventsList extends window.hassMixins.EventsMixin(PolymerElement) {
+/*
+ * @appliesMixin EventsMixin
+ */
+class EventsList extends EventsMixin(PolymerElement) {
   static get template() {
     return html`
     <style>
