--- src/util/hass-util.js
@@ -1,2 +1,0 @@
-import '..//util.js';
-
